requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
alembic==1.12.1
python-dotenv==1.0.0
pytest==7.4.3
psycopg2-binary==2.9.9
sqlacodegen==3.0.0rc3
factory-boy==3.3.0
pytest-asyncio==0.21.1
python-multipart==0.0.6
pydantic==2.5.2
asyncpg==0.29.0
gunicorn==21.2.0
httpx==0.26.0
python-jose[cryptography]
passlib[bcrypt]
